Updated JavaScript and ReadMe
The navbar and section styling have received updates. The navbar is now shown when a user scrolls, and it disappears after a user has stopped scrolling for 5 seconds. Active styling is added to a section if the section is fully in the viewport or larger than the viewport (accommodates smaller screens).